pyproject.toml
[project]
name = "procex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fernando Pérez-García", email = "fperezgarcia@microsoft.com" }
]
requires-python = ">=3.12"
dependencies = [
    "matplotlib",
    "numpy",
    "pillow",
    "scikit-learn",
    "simpleitk>=2",
    "typer",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipykernel",
    "pyright",
    "ruff",
]

[tool.ruff.lint]
select = ["ALL"]

[project.scripts]
procex = "procex.__main__:app"
